requirements.txt
fastapi
uvicorn
redis
prometheus_client
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-redis